Feat: [영상학습/면접] 유튜브 데이터 및 자체 데이터 분리 및 폴더 구조 일부 변경
