+++ lib/features/contents/repositories/entities/interface/contents_author_entity.dart
